refactor(tests): Split updateTaskById tests
